Moved classes from api into repository module
