+++ fhg-repository/src/main/java/de/jbamberger/fhg/repository/FeedConverterFactory.kt
@@ -1,4 +1,4 @@
-package de.jbamberger.api
+package de.jbamberger.fhg.repository
 
 import com.squareup.moshi.Moshi
 import com.squareup.moshi.Types
