+++ fhg-repository/src/main/java/de/jbamberger/fhg/repository/FhgApiComponent.kt
@@ -1,4 +1,4 @@
-package de.jbamberger.api
+package de.jbamberger.fhg.repository
 
 import android.app.Application
 import dagger.BindsInstance
