+++ fhg-repository/src/main/java/de/jbamberger/fhg/repository/LiveDataCallAdapter.kt
@@ -1,4 +1,4 @@
-package de.jbamberger.api
+package de.jbamberger.fhg.repository
 
 
 import android.arch.lifecycle.LiveData
