+++ fhg-repository/src/main/java/de/jbamberger/fhg/repository/data/FeedItem.kt
