+++ fhg-repository/src/main/java/de/jbamberger/fhg/repository/data/RssFeed.kt
