+++ fhg-repository/src/test/java/de/jbamberger/fhg/repository/VPlanParserTest.java
@@ -1,4 +1,4 @@
-package de.jbamberger.api;
+package de.jbamberger.fhg.repository;
 
 import org.junit.Test;
